chore(Timeline): fix timeline types [YTFRONT-4695]
